Make pre_checkin.sh sript more usable
With this commit pre_checkin.sh script does no create the changes
file by default, so it can easily be used to recreate the KIWI
files while testing and debugging without requiring to commit
anything. Also it only recreates the KIWI files for the given
product (caasp|kubic), kubic is the default if none is given.