pre_checkin: Allow to define extra bootstrap packages
This change allows to define packages section of type bootstrap per
product (kubic, caasp) by creating a file named
`${product}-extra-bootstrap-packages` (i.e. `kubic-extra-packages` or
`caasp-extra-packages`). Kiwi template, in order to support that
functionality, has to contain `_EXTRA_BOOTSTRAP_PACKAGES_` marker.

The motivation behind it is the fact that some openSUSE packages require
providers of product_flavor(openSUSE). The common solution for that is
to install i.e. openSUSE-release-ftp as a bootstrap package. However,
that problem and solution are valid only in openSUSE, not in SLE.